+++ src/acceleration/directx_backend_full.cpp
@@ -400,7 +400,7 @@ public:
             queries, numQueries, dim, vectors, numVectors, useL2);
 
         // Guard against an unexpectedly sized result from computeDistances
-        if (distances.size() != numQueries * numVectors) {
+        if (static_cast<int>(distances.size()) != numQueries * numVectors) {
             throw std::runtime_error(
                 "[DirectX] batchKnnSearch: computeDistances returned unexpected size "
                 "(expected " + std::to_string(numQueries * numVectors) +
